+++ templates/quiz/next-question.php
@@ -2,7 +2,7 @@
 /**
  * Single Quiz: Next Question button
  * @since    1.0.0
- * @version  3.9.0
+ * @version  3.14.9
  */
 
 if ( ! defined( 'ABSPATH' ) ) { exit; }
@@ -17,21 +17,13 @@ $student = llms_get_student( null );
 $attempt = $student->quizzes()->get_current_attempt( $args['quiz_id'] );
 
 $btn_text = __( 'Next Question','lifterlms' );
-
-$questions = $attempt->get( 'questions' );
-
-foreach ( $questions as $key => $value ) {
-
-	if ( $value['id'] != $question->id ) {
-		continue;
-	}
-
-	if ( $key++ > ( count( $questions ) - 1 ) ) {
-		$btn_text = __( 'Complete Quiz','lifterlms' );
-	}
+$complete = '';
+if ( $attempt->get_question_order( $args['question_id'] ) === $attempt->get_count( 'questions' ) ) {
+	 $btn_text = __( 'Complete Quiz','lifterlms' );
+	 $complete = ' llms-button-quiz-complete';
 }
 ?>
 
-<input id="llms_answer_question" type="submit" class="button llms-button-action" name="llms_answer_question" value="<?php echo $btn_text; ?>" />
+<input id="llms_answer_question" type="submit" class="button llms-button-action<?php echo $complete; ?>" name="llms_answer_question" value="<?php echo $btn_text; ?>" />
 <input type="hidden" name="action" value="llms_answer_question" />
 <?php wp_nonce_field( 'llms_answer_question' ); ?>