+++ uninstall.php
@@ -24,6 +24,7 @@ wp_clear_scheduled_hook( 'lifterlms_engagement_send_email' );
 if ( defined( 'LLMS_REMOVE_ALL_DATA' ) && true === LLMS_REMOVE_ALL_DATA ) {
 
 	include_once( dirname( __FILE__ ) . '/includes/class.llms.roles.php' );
+	include_once( dirname( __FILE__ ) . '/includes/class.llms.post-types.php' );
 
 	global $wpdb, $wp_version;